// frontend/src/gen/ChatSounds.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { QuietHours } from "./QuietHours";

export type ChatSounds = { muted: boolean, quietHours: QuietHours | null, timeZone: string, 
/**
 * Meeting and out-of-office windows, as inclusive-start/exclusive-end epoch seconds.
 */
quietWindows: Array<[number, number]>, };